refactor serve
