+++ .gitignore
@@ -1,5 +1,5 @@
 node_modules/
-cypress/screenshots
+**/cypress/screenshots
 **/cypress/videos
 build
